Docker-compose created to run server and scrapper
