--- dash.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-/home/cpt/Documents/Github/Stock-Dashbaord/venv/bin/python /home/cpt/Documents/Github/Stock-Dashbaord/dashboard.py > /home/cpt/Documents/Github/Stock-Dashbaord/test.log
