--- fetch.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-#!/bin/sh
-
-/home/cpt/Documents/Github/Stock-Dashbaord/venv/bin/python /home/cpt/Documents/Github/Stock-Dashbaord/stock_alert.py > /home/cpt/Documents/Github/Stock-Dashbaord/test.log
