+++ scripts/scrapper.sh
@@ -1,0 +1,3 @@
+cd /app
+
+exec python stock_scrapper.py
