+++ scripts/server.sh
@@ -1,0 +1,3 @@
+cd /app
+
+exec python dash_server.py
