+++ src/config.json
@@ -1,0 +1,1 @@
+[{"stock_name": "Vodafone Idea Ltd", "symbol": "IDEA:NSE", "min_price": 8.5, "max_price": 11.5, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Steel Authority of India Limited", "symbol": "SAIL:NSE", "min_price": 77.9, "max_price": 130, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Birlasoft Ltd", "symbol": "BSOFT:NSE", "min_price": 283.0, "max_price": 500, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Trident Ltd", "symbol": "TRIDENT:NSE", "min_price": 36.25, "max_price": 70, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "KPIT Technologies Ltd", "symbol": "KPITTECH:NSE", "min_price": 505.5, "max_price": 700, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Jubilant Ingrevia Ltd", "symbol": "JUBLINGREA:NSE", "min_price": 450.0, "max_price": 700, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Infosys Ltd", "symbol": "INFY:NSE", "min_price": 1459.0, "max_price": 2000, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Tata Power Company Limited", "symbol": "TATAPOWER:NSE", "min_price": 216.3, "max_price": 241.1, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Tata Motors Limited", "symbol": "TATAMOTORS:NSE", "min_price": 396.25, "max_price": 510, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Coforge Ltd", "symbol": "COFORGE:NSE", "min_price": 3754.9, "max_price": 4759, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Bharti Airtel Limited", "symbol": "BHARTIARTL:NSE", "min_price": 686.45, "max_price": 766.5, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Coal India Ltd", "symbol": "COALINDIA:NSE", "min_price": 157.2, "max_price": 236.35, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Hindustan Copper Ltd.", "symbol": "HINDCOPPER:NSE", "min_price": 101.7, "max_price": 150, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Indian Oil Corporation Ltd", "symbol": "IOC:NSE", "min_price": 66.45, "max_price": 140, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Oil & Natural Gas Corporation Limited", "symbol": "ONGC:NSE", "min_price": 127.65, "max_price": 171.35, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Reliance Industries Ltd", "symbol": "RELIANCE:NSE", "min_price": 2293.8, "max_price": 2628.85, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Affle (India) Ltd", "symbol": "AFFLE:NSE", "min_price": 1053.0, "max_price": 1300, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Borosil Renewables Ltd", "symbol": "BORORENEW:NSE", "min_price": 543.8, "max_price": 700, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Gail India", "symbol": "GAIL:NSE", "min_price": 85.2, "max_price": 155.8, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Zomato Ltd", "symbol": "ZOMATO:NSE", "min_price": 58.5, "max_price": 100, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Yes Bank Limited", "symbol": "YESBANK:NSE", "min_price": 13.15, "max_price": 15.75, "buy": true, "market": "IN", "currency": "INR"}, {"stock_name": "Larsen & Toubro Limited", "symbol": "LT:NSE", "min_price": 1610.0, "max_price": 1915.0, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "NTPC Limited", "symbol": "NTPC:NSE", "min_price": 128.95, "max_price": 165.45, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "ITC Ltd", "symbol": "ITC:NSE", "min_price": 214.75, "max_price": 332.1, "buy": false, "market": "IN", "currency": "INR"}, {"stock_name": "Apple Inc", "symbol": "AAPL:NASDAQ", "min_price": 137.59, "max_price": 180, "buy": true, "market": "US", "currency": "Dollar"}, {"stock_name": "Tesla Inc", "symbol": "TSLA:NASDAQ", "min_price": 204.99, "max_price": 900, "buy": true, "market": "US", "currency": "Dollar"}, {"stock_name": "Amazon Inc", "symbol": "AMZN:NASDAQ", "min_price": 106.9, "max_price": 3200, "buy": true, "market": "US", "currency": "Dollar"}, {"stock_name": "Meta Platform(Facebook)", "symbol": "FB2A:ETR", "min_price": 131.4, "max_price": 190, "buy": true, "market": "US", "currency": "Dollar"}, {"stock_name": "Alphabet Inc", "symbol": "GOOGL:NASDAQ", "min_price": 96.56, "max_price": 3000, "buy": true, "market": "US", "currency": "Dollar"}, {"stock_name": "NVIDIA Corporation", "symbol": "NVDA:NASDAQ", "min_price": 112.27, "max_price": 300, "buy": true, "market": "US", "currency": "Dollar"}]
